Crate flux_middle

This crate contains common type definitions that are used by other crates.

Modules

big_int

cstore

fhir

Flux High-Level Intermediate Repesentation

global_env

pretty

queries

rty

Defines how flux represents refinement types internally. Definitions in this module are used during refinement type checking. A couple of important differences between definitions in this module and in crate::fhir are:

sort_of 🔒

Macros

_Bool

_Int

_Ref

_Uint

_define_scoped

_format_args_cx

_impl_debug_with_default_cx

_join

_parens

_w

_with_cx

query_bug

try_alloc_slice

walk_list

Structs

PlaceTy

ResolverOutput

Specs

TheoryFunc

Enums

MaybeExternId

This enum serves as a type-level reminder that local ids can refer to extern specs. The abstraction is not infallible, so one shoud still be careful and decide in each situation whether to use the local id or the resolved id. Although the construction of MaybeExternId is not encapsulated, it is recommened to use GlobalEnv::maybe_extern_id to create one.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

THEORY_FUNCS

Traits

PlaceExt

Functions

def_id_to_string