flux_middle::def_id

Enum MaybeExternId

pub enum MaybeExternId<Id = LocalDefId> {
    Local(Id),
    Extern(Id, DefId),
}

This enum serves as a type-level reminder that a local definition may be a wrapper for an extern spec. This abstraction is not infallible, so one should be careful and decide in each situation whether to use the local id or the resolved id.

The construction of MaybeExternId is not encapsulated, but it is recommended to use GlobalEnv::maybe_extern_id to create one.

The enum is generic on the local Id as we use it with various kinds of local ids, e.g., [LocalDefId], [DefId], โฆ

Variants

Local(Id)

An id for a local spec.

Extern(Id, DefId)

A โdummyโ local definition wrapping an external spec. The Id is the local id of the definition corresponding to the extern spec. The DefId is the resolved id for the external definition.

Implementations

impl<Id> MaybeExternId<Id>

pub fn map<R>(self, f: impl FnOnce(Id) -> R) -> MaybeExternId<R>

pub fn local_id(self) -> Id

pub fn expect_local(self) -> Id

pub fn is_local(self) -> bool

Returns true if the maybe extern id is Local.

pub fn is_extern(&self) -> bool

Returns true if the maybe extern id is Extern.

pub fn as_local(self) -> Option<Id>

pub fn as_extern(self) -> Option<DefId>

impl<Id: Into<DefId>> MaybeExternId<Id>

pub fn resolved_id(self) -> DefId

Returns the [DefId] this id truly corresponds to, i.e, returns the [DefId] of the extern definition if Extern or converts the local id into a [DefId] if Local.