flux_middle::fhir

Enum Res

pub enum Res<Id = !> {
    Def(DefKind, DefId),
    PrimTy(PrimTy),
    SelfTyAlias {
        alias_to: DefId,
        is_trait_impl: bool,
    },
    SelfTyParam {
        trait_: DefId,
    },
    Param(ParamKind, Id),
    GlobalFunc(SpecFuncKind),
    Err,
}

The resolution of a path

The enum contains a subset of the variants in [rustc_hir::def::Res] plus some extra variants for stuff refinements resolve to.

Variants

Def(DefKind, DefId)

See [rustc_hir::def::Res::Def]

PrimTy(PrimTy)

See [rustc_hir::def::Res::PrimTy]

SelfTyAlias

See [rustc_hir::def::Res::SelfTyAlias]

Fields

alias_to: DefId

is_trait_impl: bool

SelfTyParam

See [rustc_hir::def::Res::SelfTyParam]

Fields

trait_: DefId

Param(ParamKind, Id)

A refinement parameter, e.g., declared with @n syntax

GlobalFunc(SpecFuncKind)

A refinement function defined with flux::defs! { ... }

Err

Implementations

impl<Id> Res<Id>

pub fn descr(&self) -> &'static str

pub fn is_box(&self, tcx: TyCtxt<'_>) -> bool

pub fn ns(&self) -> Option<Namespace>

Returns None if this is Res::Err

pub fn matches_ns(&self, ns: Namespace) -> bool

Always returns true if self is Res::Err

pub fn map_param_id<R>(self, f: impl FnOnce(Id) -> R) -> Res<R>

pub fn expect_param(self) -> (ParamKind, Id)