Enum flux_middle::fhir::TyKind

pub enum TyKind {
    BaseTy(BaseTy),
    Indexed(BaseTy, RefineArg),
    Exists(Vec<RefineParam>, Box<Ty>),
    Constr(Expr, Box<Ty>),
    Ptr(Lifetime, Ident),
    Ref(Lifetime, MutTy),
    Tuple(Vec<Ty>),
    Array(Box<Ty>, ArrayLen),
    RawPtr(Box<Ty>, Mutability),
    OpaqueDef(ItemId, Vec<GenericArg>, bool),
    Never,
    Hole,
}

Variants

BaseTy(BaseTy)

As a base type bty without any refinements is equivalent to bty{vs : true} we don’t technically need this variant, but we keep it around to simplify desugaring.

Indexed(BaseTy, RefineArg)

Exists(Vec<RefineParam>, Box<Ty>)

Constr(Expr, Box<Ty>)

Constrained types {T | p} are like existentials but without binders, and are useful for specifying constraints on indexed values e.g. {i32[@a] | 0 <= a}

Ptr(Lifetime, Ident)

Ref(Lifetime, MutTy)

Tuple(Vec<Ty>)

Array(Box<Ty>, ArrayLen)

RawPtr(Box<Ty>, Mutability)

OpaqueDef(ItemId, Vec<GenericArg>, bool)

Never

Hole