Struct flux_middle::fhir::Map

pub struct Map {

    generics: FrozenMap<LocalDefId, Box<Generics>, BuildHasherDefault<FxHasher>>,
    predicates: ItemPredicates,
    opaque_tys: UnordMap<LocalDefId, OpaqueTy>,
    func_decls: FxHashMap<Symbol, FuncDecl>,
    sort_decls: SortDecls,
    flux_items: FxHashMap<Symbol, FluxItem>,
    consts: FxHashMap<Symbol, ConstInfo>,
    refined_by: UnordMap<LocalDefId, RefinedBy>,
    type_aliases: FxHashMap<LocalDefId, TyAlias>,
    structs: FxHashMap<LocalDefId, StructDef>,
    enums: FxHashMap<LocalDefId, EnumDef>,
    fns: FxHashMap<LocalDefId, FnSig>,
    fn_quals: FxHashMap<LocalDefId, Vec<SurfaceIdent>>,
    trusted: UnordSet<LocalDefId>,
    externs: UnordMap<DefId, LocalDefId>,

}

A map between rust definitions and flux annotations in their desugared fhir form.

note: Map is a very generic name, so we typically use the type qualified as fhir::Map.

Fields
generics: FrozenMap<LocalDefId, Box<Generics>, BuildHasherDefault<FxHasher>>predicates: ItemPredicatesopaque_tys: UnordMap<LocalDefId, OpaqueTy>func_decls: FxHashMap<Symbol, FuncDecl>sort_decls: SortDeclsflux_items: FxHashMap<Symbol, FluxItem>consts: FxHashMap<Symbol, ConstInfo>refined_by: UnordMap<LocalDefId, RefinedBy>type_aliases: FxHashMap<LocalDefId, TyAlias>structs: FxHashMap<LocalDefId, StructDef>enums: FxHashMap<LocalDefId, EnumDef>fns: FxHashMap<LocalDefId, FnSig>fn_quals: FxHashMap<LocalDefId, Vec<SurfaceIdent>>trusted: UnordSet<LocalDefId>externs: UnordMap<DefId, LocalDefId>
Implementations

impl Map

pub fn new() -> Self

pub fn insert_generics(&self, def_id: LocalDefId, generics: Generics)

pub fn insert_generic_predicates( &mut self, def_id: LocalDefId, predicates: GenericPredicates )

pub fn insert_opaque_tys(&mut self, opaque_tys: UnordMap<LocalDefId, OpaqueTy>)

pub fn get_generics(&self, def_id: LocalDefId) -> Option<&Generics>

pub fn get_refine_params( &self, tcx: TyCtxt<'_>, def_id: LocalDefId ) -> Option<&[RefineParam]>

pub fn get_generic_predicates( &self, def_id: LocalDefId ) -> Option<&GenericPredicates>

pub fn get_opaque_ty(&self, def_id: LocalDefId) -> Option<&OpaqueTy>

pub fn insert_qualifier(&mut self, qualifier: Qualifier)

pub fn qualifiers(&self) -> impl Iterator<Item = &Qualifier>

pub fn insert_fn_sig(&mut self, def_id: LocalDefId, fn_sig: FnSig)

pub fn insert_fn_quals(&mut self, def_id: LocalDefId, quals: Vec<SurfaceIdent>)

pub fn add_trusted(&mut self, def_id: LocalDefId)

pub fn fn_sigs(&self) -> impl Iterator<Item = (LocalDefId, &FnSig)>

pub fn get_fn_sig(&self, def_id: LocalDefId) -> &FnSig

pub fn fn_quals(&self) -> impl Iterator<Item = (LocalDefId, &Vec<SurfaceIdent>)>

pub fn get_fn_quals( &self, def_id: LocalDefId ) -> impl Iterator<Item = SurfaceIdent> + '_

pub fn is_trusted(&self, def_id: LocalDefId) -> bool

pub fn insert_extern(&mut self, extern_def_id: DefId, local_def_id: LocalDefId)

pub fn get_extern(&self, extern_def_id: DefId) -> Option<LocalDefId>

pub fn insert_refined_by(&mut self, def_id: LocalDefId, refined_by: RefinedBy)

pub fn refined_by(&self, def_id: LocalDefId) -> &RefinedBy

pub fn insert_type_alias(&mut self, def_id: LocalDefId, alias: TyAlias)

pub fn type_aliases(&self) -> impl Iterator<Item = &TyAlias>

pub fn get_type_alias(&self, def_id: impl Borrow<LocalDefId>) -> &TyAlias

pub fn insert_struct(&mut self, def_id: LocalDefId, struct_def: StructDef)

pub fn structs(&self) -> impl Iterator<Item = &StructDef>

pub fn get_struct(&self, def_id: impl Borrow<LocalDefId>) -> &StructDef

pub fn insert_enum(&mut self, def_id: LocalDefId, enum_def: EnumDef)

pub fn enums(&self) -> impl Iterator<Item = &EnumDef>

pub fn get_enum(&self, def_id: impl Borrow<LocalDefId>) -> &EnumDef

pub fn insert_const(&mut self, c: ConstInfo)

pub fn consts(&self) -> impl Iterator<Item = &ConstInfo>

pub fn const_by_name(&self, name: impl Borrow<Symbol>) -> Option<&ConstInfo>

fn insert_theory_func( &mut self, name: Symbol, fixpoint_name: Symbol, params: usize, inputs: Vec<Sort>, output: Sort )

fn insert_theory_funcs(&mut self)

pub fn insert_func_decl(&mut self, symb: Symbol, uif: FuncDecl)

pub fn func_decls(&self) -> impl Iterator<Item = &FuncDecl>

pub fn func_decl(&self, sym: impl Borrow<Symbol>) -> Option<&FuncDecl>

pub fn insert_defn(&mut self, symb: Symbol, defn: Defn)

pub fn defns(&self) -> impl Iterator<Item = &Defn>

pub fn defn(&self, sym: impl Borrow<Symbol>) -> Option<&Defn>

pub fn insert_sort_decl(&mut self, sort_decl: SortDecl)

pub fn sort_decls(&self) -> &SortDecls

pub fn sort_decl(&self, name: impl Borrow<Symbol>) -> Option<&SortDecl>

pub fn get_flux_item(&self, name: impl Borrow<Symbol>) -> Option<&FluxItem>