flux_middle::rty
Enum BaseTy

pub enum BaseTy {

    Int(IntTy),
    Uint(UintTy),
    Bool,
    Str,
    Char,
    Slice(Ty),
    Adt(AdtDef, GenericArgs),
    Float(FloatTy),
    RawPtr(Ty, Mutability),
    Ref(Region, Ty, Mutability),
    FnPtr(PolyFnSig),
    FnDef(DefId, GenericArgs),
    Tuple(List<Ty>),
    Alias(AliasKind, AliasTy),
    Array(Ty, Const),
    Never,
    Closure(DefId, List<Ty>, GenericArgs),
    Coroutine(DefId, Ty, List<Ty>),
    Dynamic(List<Binder<ExistentialPredicate>>, Region),
    Param(ParamTy),
    Infer(TyVid),

}

Variants

Int(IntTy)

Uint(UintTy)

Bool

Str

Char

Slice(Ty)

Adt(AdtDef, GenericArgs)

Float(FloatTy)

RawPtr(Ty, Mutability)

Ref(Region, Ty, Mutability)

FnPtr(PolyFnSig)

FnDef(DefId, GenericArgs)

Tuple(List<Ty>)

Alias(AliasKind, AliasTy)

Array(Ty, Const)

Never

Closure(DefId, List<Ty>, GenericArgs)

Coroutine(DefId, Ty, List<Ty>)

Dynamic(List<Binder<ExistentialPredicate>>, Region)

Param(ParamTy)

Infer(TyVid)

Implementations

impl BaseTy

pub fn opaque(alias_ty: AliasTy) -> BaseTy

pub fn projection(alias_ty: AliasTy) -> BaseTy

pub fn adt(adt_def: AdtDef, args: impl Into<GenericArgs>) -> BaseTy

pub fn fn_def(def_id: DefId, args: impl Into<GenericArgs>) -> BaseTy

pub fn from_primitive_str(s: &str) -> Option<BaseTy>

pub fn primitive_symbol(&self) -> Option<Symbol>

If self is a primitive, return its Symbol.

pub fn is_integral(&self) -> bool

pub fn is_numeric(&self) -> bool

pub fn is_signed(&self) -> bool

pub fn is_unsigned(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_bool(&self) -> bool

fn is_struct(&self) -> bool

fn is_array(&self) -> bool

fn is_slice(&self) -> bool

fn is_adt(&self) -> bool

pub fn is_box(&self) -> bool

pub fn is_char(&self) -> bool

pub fn is_str(&self) -> bool

pub fn unpack_box(&self) -> Option<(&Ty, &Ty)>

pub fn invariants(&self, overflow_checking: bool) -> &[Invariant]

pub fn to_ty(&self) -> Ty

pub fn to_subset_ty_ctor(&self) -> SubsetTyCtor

pub fn expect_adt(&self) -> (&AdtDef, &[GenericArg])

pub fn is_atom(&self) -> bool

A type is an atom if it is “self-delimiting”, i.e., it has a clear boundary when printed. This is used to avoid unnecessary parenthesis when pretty printing.

impl BaseTy

pub fn sort(&self) -> Sort