Enum flux_middle::rty::ExprKind

pub enum ExprKind {

    Var(Var),
    Local(Local),
    Constant(Constant),
    ConstDefId(DefId),
    BinaryOp(BinOp, Expr, Expr),
    App(Expr, List<Expr>),
    GlobalFunc(Symbol, FuncKind),
    UnaryOp(UnOp, Expr),
    TupleProj(Expr, u32),
    Tuple(List<Expr>),
    PathProj(Expr, FieldIdx),
    IfThenElse(Expr, Expr, Expr),
    KVar(KVar),
    Abs(Binder<Expr>),
    Hole,

}

Variants

Var(Var)

Local(Local)

Constant(Constant)

ConstDefId(DefId)

BinaryOp(BinOp, Expr, Expr)

App(Expr, List<Expr>)

GlobalFunc(Symbol, FuncKind)

UnaryOp(UnOp, Expr)

TupleProj(Expr, u32)

Tuple(List<Expr>)

PathProj(Expr, FieldIdx)

IfThenElse(Expr, Expr, Expr)

KVar(KVar)

Abs(Binder<Expr>)

Lambda abstractions. They are purely syntactic and we don’t encode them in the logic. As such, they have some syntactic restrictions that we must carefully maintain:

They can appear as an index at the top level.

We can only substitute an abstraction for a variable in function position (or as an index). More generaly, we need to partially evaluate expressions such that all abstractions in non-index position are eliminated before encoding into fixpoint. Right now, the implementation only evaluates abstractions that are immediately applied to arguments, thus the restriction.

Hole

Implementations

impl ExprKind

fn intern_at(self, espan: Option<ESpan>) -> Expr

fn intern(self) -> Expr