Module flux_middle::rty::evars

Modules

pretty 🔒

Structs

EVar

An existential varriable is identified by a context and an id. Two evars are considered equal if both the context and id are equal.

EVarCtxt 🔒

EVarCxId

EVarGen

A generator of evars. Evars are associated with a context. Contexts are visited in a stack-like fashion. When entering a context data can be associated to that particular context. When a context is exited, all its evars are put in a list of pending contexts. No more evars can be generated for pending contexts, but they can still be unified. At any point, pending contexts can be solved, returning a solution for all of their evars or an error if some evar hasn’t been unifed yet.

EVarSol

EVid 🔒

Existential variable id

UnsolvedEvar

Enums

EVarState 🔒

Statics

NEXT_CTXT_ID 🔒