Struct flux_middle::rty::evars::EVarGen

pub struct EVarGen<T> {
    stack: FxIndexMap<EVarCxId, EVarCtxt<T>>,
    pending: FxIndexMap<EVarCxId, EVarCtxt<T>>,
}

A generator of evars. Evars are associated with a context. Contexts are visited in a stack-like fashion. When entering a context data can be associated to that particular context. When a context is exited, all its evars are put in a list of pending contexts. No more evars can be generated for pending contexts, but they can still be unified. At any point, pending contexts can be solved, returning a solution for all of their evars or an error if some evar hasn’t been unifed yet.

Fields
stack: FxIndexMap<EVarCxId, EVarCtxt<T>>pending: FxIndexMap<EVarCxId, EVarCtxt<T>>
Implementations

impl<T> EVarGen<T>

pub fn enter_context(&mut self, data: T) -> EVarCxId

Enters a new context generating a context id guaranteed to be globally fresh. That is, it will be unique among context ids generated by this or other instances of EVarGen. The context can be used to generate fresh evars by calling EVarGen::fresh_in.

pub fn exit_context(&mut self) -> EVarCxId

Exit the current context, putting it in the pending list.

pub fn fresh_in(&mut self, cxid: EVarCxId) -> EVar

Generates an evar guaranteed to be fresh in the provided context.

Panics

This function panics if the context cxid was not generated by this EVarGen instance or has already been popped.

pub fn fresh_in_current(&mut self) -> EVar

Generates a fresh evar in the current (top of the stack) context

pub fn data(&self, cxid: EVarCxId) -> &T

Returns the data associated with the context cxid

Panics

This function panics if the context cxid was not generated by this EVarGen instance.

pub fn unify(&mut self, evar: EVar, arg: impl Into<Expr>, replace: bool)

pub fn try_solve_pending(&mut self) -> Result<EVarSol, UnsolvedEvar>

Try to solve evars in all pending contexts and then empty the pending list.