flux_middle::rty::refining

Function make_vars_and_sorts_from_bound_vars

fn make_vars_and_sorts_from_bound_vars<'a, I, II>(vars: I) -> Vec<(Var, Sort)>
where
    I: IntoIterator<IntoIter = II>,
    II: DoubleEndedIterator<Item = &'a BoundVariableKind>,

NOTE(CK):

Skips params (we don’t presently handle polymorphism, though even if we did, I’m not sure that we need to pass params to the weak kvars).

Skips locs because we can’t encode those.

Skips unit + unit adts because they otherwise get encoded as a 0 tuple to fixpoint because we use them in the args to a weak kvar, which we don’t want to do.