flux_middle::rty
Module region_matching

Structs

RegionSubst 🔒

Functions

replace_regions_with_unique_vars 🔒

Replace all non-bound regions with a rty::ReVar assigning each a unique rty::RegionVid. This is used to have a unique identifier for each position such that we can infer a region substitution.

rty_match_regions

ty_match_regions

See flux_refineck::type_env::TypeEnv::assign