flux_middle::rty

Struct KVar

pub struct KVar {
    pub kvid: KVid,
    pub self_args: usize,
    pub args: List<Expr>,
}

In theory a kvar is just an unknown predicate that can use some variables in scope. In practice, fixpoint makes a difference between the first and the rest of the arguments, the first one being the kvarโs self argument. Fixpoint will only instantiate qualifiers that use the self argument. Flux generalizes the self argument to be a list. We call the rest of the arguments the scope.

Fields
kvid: KVidself_args: usize

The number of arguments consider to be self arguments.

args: List<Expr>

The list of all arguments with the self arguments at the beginning, i.e., the list of self arguments followed by the scope.

Implementations

impl KVar

pub fn new(kvid: KVid, self_args: usize, args: Vec<Expr>) -> Self

fn self_args(&self) -> &[Expr]

fn scope(&self) -> &[Expr]