flux_middle::rty

Struct Ty

pub struct Ty(Interned<TyKind>);

Tuple Fields
0: Interned<TyKind>
Implementations

impl Ty

pub fn shallow_canonicalize(&self) -> CanonicalTy

Hoist existentials and predicates inside the type stopping when encountering the first type constructor.

impl Ty

pub fn kind(&self) -> &TyKind

pub fn trait_object_dummy_self() -> Ty

Dummy type used for the Self of a TraitRef created when converting a trait object, and which gets removed in ExistentialTraitRef. This type must not appear anywhere in other converted types and must be a valid rustc type (i.e., we must be able to call to_rustc on it). TyKind::Infer(TyVid(0)) does the job, with the caveat that we must skip 0 when generating TyKind::Infer for “type holes”.

pub fn dynamic( preds: impl Into<List<Binder<ExistentialPredicate>>>, region: Region, ) -> Ty

pub fn strg_ref(re: Region, path: Path, ty: Ty) -> Ty

pub fn ptr(pk: impl Into<PtrKind>, path: impl Into<Path>) -> Ty

pub fn constr(p: impl Into<Expr>, ty: Ty) -> Ty

pub fn uninit() -> Ty

pub fn indexed(bty: BaseTy, idx: impl Into<Expr>) -> Ty

pub fn exists(ty: Binder<Ty>) -> Ty

pub fn exists_with_constr(bty: BaseTy, pred: Expr) -> Ty

pub fn discr(adt_def: AdtDef, place: Place) -> Ty

pub fn unit() -> Ty

pub fn bool() -> Ty

pub fn int(int_ty: IntTy) -> Ty

pub fn uint(uint_ty: UintTy) -> Ty

pub fn param(param_ty: ParamTy) -> Ty

pub fn downcast( adt: AdtDef, args: GenericArgs, ty: Ty, variant: VariantIdx, fields: List<Ty>, ) -> Ty

pub fn blocked(ty: Ty) -> Ty

pub fn str() -> Ty

pub fn char() -> Ty

pub fn float(float_ty: FloatTy) -> Ty

pub fn mk_ref(region: Region, ty: Ty, mutbl: Mutability) -> Ty

pub fn mk_slice(ty: Ty) -> Ty

pub fn mk_box( genv: GlobalEnv<'_, '_>, deref_ty: Ty, alloc_ty: GenericArg, ) -> QueryResult<Ty>

pub fn mk_box_with_default_alloc( genv: GlobalEnv<'_, '_>, deref_ty: Ty, ) -> QueryResult<Ty>

pub fn tuple(tys: impl Into<List<Ty>>) -> Ty

pub fn array(ty: Ty, c: Const) -> Ty

pub fn closure( did: DefId, tys: impl Into<List<Ty>>, args: &GenericArgs, no_panic: bool, ) -> Ty

pub fn coroutine( did: DefId, resume_ty: Ty, upvar_tys: List<Ty>, args: GenericArgs, ) -> Ty

pub fn never() -> Ty

pub fn infer(vid: TyVid) -> Ty

pub fn unconstr(&self) -> (Ty, Expr)

pub fn unblocked(&self) -> Ty

pub fn is_integral(&self) -> bool

Whether the type is an int or a uint

pub fn is_bool(&self) -> bool

Whether the type is a bool

pub fn is_char(&self) -> bool

Whether the type is a char

pub fn is_uninit(&self) -> bool

pub fn is_box(&self) -> bool

pub fn is_struct(&self) -> bool

pub fn is_array(&self) -> bool

pub fn is_slice(&self) -> bool

pub fn as_bty_skipping_existentials(&self) -> Option<&BaseTy>

pub fn expect_discr(&self) -> (&AdtDef, &Place)

pub fn expect_adt(&self) -> (&AdtDef, &[GenericArg], &Expr)

pub fn expect_tuple(&self) -> &[Ty]

pub fn simplify_type(&self) -> Option<SimplifiedType>