Type Alias flux_middle::rty::Expr

pub type Expr = Interned<ExprS>;

Aliased Type

struct Expr {
    arc: Arc<ExprS, Global>,
}

Fields
arc: Arc<ExprS, Global>
Implementations

impl Expr

pub fn at_base(self, base: Option<ESpan>) -> Expr

pub fn span(&self) -> Option<ESpan>

pub fn tt() -> Expr

pub fn ff() -> Expr

pub fn and(exprs: impl IntoIterator<Item = Expr>) -> Expr

pub fn or(exprs: impl IntoIterator<Item = Expr>) -> Expr

pub fn zero() -> Expr

pub fn one() -> Expr

pub fn int_max(int_ty: IntTy) -> Expr

pub fn int_min(int_ty: IntTy) -> Expr

pub fn uint_max(uint_ty: UintTy) -> Expr

pub fn nu() -> Expr

pub fn as_tuple(&self) -> &[Expr]

pub fn expect_tuple(&self) -> &[Expr]

pub fn unit() -> Expr

pub fn var(var: Var, espan: Option<ESpan>) -> Expr

pub fn fvar(name: Name) -> Expr

pub fn evar(evar: EVar) -> Expr

pub fn late_bvar(bvar: DebruijnIndex, idx: u32) -> Expr

pub fn early_bvar(idx: u32) -> Expr

pub fn local(local: Local, espan: Option<ESpan>) -> Expr

pub fn constant(c: Constant) -> Expr

pub fn constant_at(c: Constant, espan: Option<ESpan>) -> Expr

pub fn const_def_id(c: DefId, espan: Option<ESpan>) -> Expr

pub fn tuple(exprs: impl Into<List<Expr>>) -> Expr

pub fn from_bits(bty: &BaseTy, bits: u128) -> Expr

pub fn ite( p: impl Into<Expr>, e1: impl Into<Expr>, e2: impl Into<Expr>, espan: Option<ESpan> ) -> Expr

pub fn abs(body: Binder<Expr>) -> Expr

pub fn hole(kind: HoleKind) -> Expr

pub fn kvar(kvar: KVar) -> Expr

pub fn binary_op( op: BinOp, e1: impl Into<Expr>, e2: impl Into<Expr>, espan: Option<ESpan> ) -> Expr

pub fn app( func: impl Into<Expr>, args: impl Into<List<Expr>>, espan: Option<ESpan> ) -> Expr

pub fn global_func(func: Symbol, kind: FuncKind) -> Expr

pub fn unary_op(op: UnOp, e: impl Into<Expr>, espan: Option<ESpan>) -> Expr

pub fn eq_at( e1: impl Into<Expr>, e2: impl Into<Expr>, espan: Option<ESpan> ) -> Expr

pub fn eq(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn ne(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn ge(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn gt(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn lt(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn le(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn implies(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn tuple_proj(e: impl Into<Expr>, proj: u32, espan: Option<ESpan>) -> Expr

pub fn tuple_projs(e: impl Into<Expr>, projs: &[u32]) -> Expr

pub fn path_proj(base: Expr, field: FieldIdx) -> Expr

pub fn not(&self) -> Expr

pub fn neg(&self) -> Expr

pub fn kind(&self) -> &ExprKind

pub fn is_atom(&self) -> bool

An expression is an atom if it is “self-delimiting”, i.e., it has a clear boundary when printed. This is used to avoid unnecesary parenthesis when pretty printing.

pub fn is_trivially_true(&self) -> bool

Simple syntactic check to see if the expression is a trivially true predicate. This is used mostly for filtering predicates when pretty printing but also to simplify the constraint before encoding it into fixpoint.

pub fn is_true(&self) -> bool

Whether the expression is literally the constant true.

pub fn is_binary_op(&self) -> bool

fn const_op(op: &BinOp, c1: &Constant, c2: &Constant) -> Option<Constant>

pub fn simplify(&self) -> Expr

Simplify the expression by removing double negations, short-circuiting boolean connectives and doing constant folding. Note that we also have TypeFoldable::normalize which applies beta reductions for tuples and abstractions.

pub fn to_var(&self) -> Option<Var>

pub fn to_loc(&self) -> Option<Loc>

pub fn to_path(&self) -> Option<Path>

pub fn is_abs(&self) -> bool

pub fn is_tuple(&self) -> bool

pub fn eta_expand_abs(&self, sorts: &[Sort]) -> Binder<Expr>

pub fn eta_expand_tuple(&self, sort: &Sort) -> Expr

pub fn fold_sort(sort: &Sort, f: impl FnMut(&Sort) -> Expr) -> Expr

impl<T: Internable> Interned<T>

pub fn new(obj: T) -> Self

impl<T: Internable + ?Sized> Interned<T>

fn select( obj: &T ) -> (RwLockWriteGuard<'static, HashMap<Arc<T>, SharedValue<()>, BuildHasherDefault<FxHasher>>>, u64)

impl<T: Internable + ?Sized> Interned<T>

fn drop_slow(&mut self)