Type Alias flux_middle::rty::GenericArgs

pub type GenericArgs = List<GenericArg>;

Aliased Type

struct GenericArgs {
    arc: Arc<[GenericArg]>,
}

Fields
arc: Arc<[GenericArg]>
Implementations

impl GenericArgs

pub fn box_args(&self) -> (&Ty, &Ty)

pub fn for_item<F>( genv: GlobalEnv<'_, '_>, def_id: DefId, mk_kind: F, ) -> QueryResult<Self>
where F: FnMut(&GenericParamDef, &[GenericArg]) -> GenericArg,

Creates a GenericArgs from the definition of generic parameters, by calling a closure to obtain arg. The closures get to observe the GenericArgs as they’re being built, which can be used to correctly replace defaults of generic parameters.

pub fn identity_for_item( genv: GlobalEnv<'_, '_>, def_id: impl Into<DefId>, ) -> QueryResult<Self>

fn fill_item<F>( genv: GlobalEnv<'_, '_>, args: &mut Vec<GenericArg>, generics: &Generics, mk_kind: &mut F, ) -> QueryResult<()>
where F: FnMut(&GenericParamDef, &[GenericArg]) -> GenericArg,

fn to_rustc<'tcx>(&self, tcx: TyCtxt<'tcx>) -> GenericArgsRef<'tcx>