flux_middle::rty

Type Alias SubsetTyCtor

pub type SubsetTyCtor = Binder<SubsetTy>;

A type constructor meant to be used as generic a argument of kind base. This is just an alias to Binder<SubsetTy>, but we expect the binder to have a single bound variable of the sort of the underlying base type.

Aliased Type

pub struct SubsetTyCtor {
    vars: Interned<[BoundVariableKind]>,
    value: SubsetTy,
}

Fields
vars: Interned<[BoundVariableKind]>value: SubsetTy
Implementations

impl SubsetTyCtor

pub fn as_bty_skipping_binder(&self) -> &BaseTy

pub fn to_ty(&self) -> Ty

pub fn to_ty_ctor(&self) -> TyCtor