Struct flux_middle::rustc::mir::FieldIdx

pub struct FieldIdx {
    pub(crate) private: u32,
}

The source-order index of a field in a variant.

This is how most code after type checking refers to fields, rather than using names (as names have hygiene complications and more complex lookup).

Particularly for repr(Rust) types, this may not be the same as layout order. (It is for repr(C) structs, however.)

For example, in the following types,

enum Demo1 {
   Variant0 { a: Never, b: i32 } = 100,
   Variant1 { c: u8, d: u64 } = 10,
}
struct Demo2 { e: u8, f: u16, g: u8 }

b is FieldIdx(1) in VariantIdx(0), d is FieldIdx(1) in VariantIdx(1), and f is FieldIdx(1) in VariantIdx(0).

Fields
private: u32