Struct flux_middle::rustc::mir::SwitchTargets

pub struct SwitchTargets {
    pub(super) values: SmallVec<[Pu128; 1]>,
    pub(super) targets: SmallVec<[BasicBlock; 2]>,
}

Fields
values: SmallVec<[Pu128; 1]>targets: SmallVec<[BasicBlock; 2]>
Implementations

impl SwitchTargets

pub fn new( targets: impl Iterator<Item = (u128, BasicBlock)>, otherwise: BasicBlock ) -> SwitchTargets

Creates switch targets from an iterator of values and target blocks.

The iterator may be empty, in which case the SwitchInt instruction is equivalent to goto otherwise;.

pub fn static_if( value: u128, then: BasicBlock, else_: BasicBlock ) -> SwitchTargets

Builds a switch targets definition that jumps to then if the tested value equals value, and to else_ if not.

pub fn as_static_if(&self) -> Option<(u128, BasicBlock, BasicBlock)>

Inverse of SwitchTargets::static_if.

pub fn otherwise(&self) -> BasicBlock

Returns the fallback target that is jumped to when none of the values match the operand.

pub fn iter(&self) -> SwitchTargetsIter<'_>

Returns an iterator over the switch targets.

The iterator will yield tuples containing the value and corresponding target to jump to, not including the otherwise fallback target.

Note that this may yield 0 elements. Only the otherwise branch is mandatory.

pub fn all_targets(&self) -> &[BasicBlock]

Returns a slice with all possible jump targets (including the fallback target).

pub fn all_targets_mut(&mut self) -> &mut [BasicBlock]

pub fn target_for_value(&self, value: u128) -> BasicBlock

Finds the BasicBlock to which this SwitchInt will branch given the specific value. This cannot fail, as it’ll return the otherwise branch if there’s not a specific match for the value.

pub fn add_target(&mut self, value: u128, bb: BasicBlock)

Adds a new target to the switch. But You cannot add an already present value.

pub fn is_distinct(&self) -> bool

Returns true if all targets (including the fallback target) are distinct.