Module flux_middle::rustc::ty

A simplified version of rust types.

Modules

subst 🔒

Structs

AdtDef

AdtDefData

AliasTy

Binder

BoundRegion

BoundVar

Clause

ClosureArgs

ClosureArgsParts

DebruijnIndex

A De Bruijn index is a standard means of representing regions (and perhaps later types) in a higher-ranked setting. In particular, imagine a type like this:

EarlyBinder

EarlyBoundRegion

FieldDef

FnSig

GeneratorArgs

GeneratorArgsParts

GenericParamDef

GenericPredicates

Generics

ParamTy

ProjectionPredicate

RegionVar

During borrow checking, rustc generates fresh region variable ids for each structurally different position in a type. For example, given a function

RegionVid

A region (lifetime) variable ID.

ScalarInt

The raw bytes of a simple value.

TraitPredicate

TraitRef

Ty

TyS 🔒

ValueConst

VariantDef

Enums

AliasKind

BoundRegionKind

BoundVariableKind

ClauseKind

Const

FloatTy

GenericArg

GenericParamDefKind

IntTy

Mutability

Region

TyKind

UintTy

Functions

region_to_string 🔒

Type Aliases

GenericArgs

PolyFnSig