flux_middle
Struct Specs

pub struct Specs {

    pub fn_sigs: UnordMap<OwnerId, FnSpec>,
    pub structs: UnordMap<OwnerId, StructDef>,
    pub traits: UnordMap<OwnerId, Trait>,
    pub impls: UnordMap<OwnerId, Impl>,
    pub enums: UnordMap<OwnerId, EnumDef>,
    pub flux_items_by_parent: FxIndexMap<OwnerId, Vec<Item>>,
    pub ty_aliases: UnordMap<OwnerId, Option<TyAlias>>,
    pub ignores: UnordMap<LocalDefId, Ignored>,
    pub trusted: UnordMap<LocalDefId, Trusted>,
    pub trusted_impl: UnordMap<LocalDefId, Trusted>,
    pub check_overflows: UnordMap<LocalDefId, CheckOverflow>,
    pub crate_config: Option<CrateConfig>,
    pub should_fail: UnordSet<LocalDefId>,
    pub(crate) dummy_extern: UnordSet<LocalDefId>,
    pub(crate) extern_id_to_local_id: UnordMap<DefId, LocalDefId>,
    pub(crate) local_id_to_extern_id: UnordMap<LocalDefId, DefId>,

}

Fields
fn_sigs: UnordMap<OwnerId, FnSpec>structs: UnordMap<OwnerId, StructDef>traits: UnordMap<OwnerId, Trait>impls: UnordMap<OwnerId, Impl>enums: UnordMap<OwnerId, EnumDef>flux_items_by_parent: FxIndexMap<OwnerId, Vec<Item>>ty_aliases: UnordMap<OwnerId, Option<TyAlias>>ignores: UnordMap<LocalDefId, Ignored>trusted: UnordMap<LocalDefId, Trusted>trusted_impl: UnordMap<LocalDefId, Trusted>check_overflows: UnordMap<LocalDefId, CheckOverflow>crate_config: Option<CrateConfig>should_fail: UnordSet<LocalDefId>dummy_extern: UnordSet<LocalDefId>

Set of dummy items generated by the extern spec macro we must completely ignore. This is not the same as ignored items because, for ignored items, we still need to return errors for queries and handle them gracefully in order to report them at the use it.

If an item is in this set, all its descendants are also consider dummy (but they may not be in the set).

extern_id_to_local_id: UnordMap<DefId, LocalDefId>local_id_to_extern_id: UnordMap<LocalDefId, DefId>
Implementations

impl Specs

pub fn insert_extern_spec_id_mapping( &mut self, local_id: LocalDefId, extern_id: DefId, ) -> Result<(), ExternSpecMappingErr>

pub fn insert_dummy(&mut self, owner_id: OwnerId)