Crate flux_middle

This crate contains common type definitions that are used by other crates.

Modules

big_int

builtin_assoc_refts 🔒

cstore

def_id

fhir

Flux High-Level Intermediate Representation

global_env

metrics

pretty

queries

rty

Defines how flux represents refinement types internally. Definitions in this module are used during refinement type checking. A couple of important differences between definitions in this module and in crate::fhir are:

sort_of 🔒

Macros

_Bool

_Char

_Int

_Ref

_Uint

_format_args_cx

_format_cx

_impl_debug_with_default_cx

_join

_parens

_w

_with_cx

query_bug

try_alloc_slice

walk_list

Structs

PlaceTy

ResolverOutput

Specs

TheoryFunc

Enums

ExternSpecMappingErr

Represents errors that can occur when inserting a mapping between a LocalDefId and a DefId for an extern spec.

FixpointQueryKind

The different reasons we issue fixpoint queries. This is used to dissambiguate queries that are issued for the same item.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

THEORY_FUNCS

Traits

PlaceExt

Functions

def_id_to_string

name_of_thy_func

sort_of_thy_func 🔒