Crate flux_middle

This crate contains common type definitions that are used by other crates.

Modules

const_eval

cstore

fhir

Flux High-Level Intermediate Repesentation

global_env

intern

Global Arc-based object interning infrastructure.

pretty

queries

rty

Defines how flux represents refinement types internally. Definitions in this module are used during refinement type checking. A couple of important differences between definitions in this module and in crate::fhir are:

rustc

This module contains simplified versions of some structures in rustc. The definitions in this module can be understood as the current supported subset of rust. As we implement more features we should be able to work directly on rustc’s structures.

sort_of 🔒

Macros

_Bool

_Int

_Ref

_Uint

_define_scoped

_format_args_cx

_impl_debug_with_default_cx

_join

_parens

_w

_with_cx

try_alloc_slice

walk_list

Structs

ResolverOutput

Specs

TheoryFunc

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

THEORY_FUNCS