Module flux_middle::fhir

Flux High-Level Intermediate Repesentation

The fhir corresponds to the desugared version of source level flux annotations. The main difference with the surface syntax is that the list of refinement parameters is explicit in fhir. For example, the following signature

fn(x: &strg i32[@n]) ensures x: i32[n + 1]

desugars to

for<n: int, l: loc> fn(l: i32[n]; ptr(l)) ensures l: i32[n + 1].

The main analysis performed on the fhir is well-formedness, thus, the fhir keeps track of spans for refinement expressions to report errors.

The name fhir is borrowed (pun intended) from rustc’s hir to refer to something a bit lower than the surface syntax.

Modules

lift

“Lift” HIR types into FHIR types.

Structs

ArrayLen

BaseTy

These are types of things that may be refined with indices or existentials

ConstInfo

Defn

EnumDef

Expr

FhirId

A unique identifier for a node in the AST. Like HirId it is composed of an owner and a local_id. We don’t generate ids for all nodes, but only for those we need to remember information elaborated during well-formedness checking to later be used during conversion into rty.

FieldDef

FnInfo

FnOutput

FnSig

FuncDecl

FuncSort

GenericParam

GenericPredicates

Generics

Ident

ItemLocalId

An ItemLocalId uniquely identifies something within a given “item-like”.

LocalTableInContext

LocalTableInContextMut

Map

A map between rust definitions and flux annotations in their desugared fhir form.

MutTy

Name

OpaqueTy

Path

PolyFuncSort

PolyTraitRef

Qualifier

RefineParam

RefinedBy

Information about the refinement parameters associated with a type alias or a struct/enum.

SortDecl

SortVid

A Sort vvariable id

StructDef

Ty

TyAlias

TypeBinding

VariantDef

VariantIdx

The source-order index of a variant in a type.

VariantRet

WfckResults

WhereBoundPredicate

Enums

BaseTyKind

BinOp

Coercion

Constraint

ExprKind

FluxItem

FluxLocalDefId

FluxOwnerId

Owner version of FluxLocalDefId

Func

FuncKind

GenericArg

GenericBound

GenericParamKind

InferMode

Inference mode for parameter at function calls

Lifetime

Our surface syntax doesn’t have lifetimes. To deal with them we create a hole for every lifetime which we then resolve during annot_check when zipping against the lifted version.

Lit

Mutability

PrimTy

Not represented directly in the AST; referred to by name through a ty_path.

QPath

RefineArg

Res

Sort

SortCtor

StructKind

TraitBoundModifier

TyKind

UnOp

WeakKind

Type Aliases

Cache 🔒

GenericBounds

ItemLocalMap

ItemPredicates

SortDecls

SurfaceIdent