Module flux_middle::fhir

Flux High-Level Intermediate Repesentation

The fhir corresponds to the desugared version of source level flux annotations. The main difference with the surface syntax is that the list of refinement parameters is explicit in fhir. For example, the following signature

fn(x: &strg i32[@n]) ensures x: i32[n + 1]

desugars to

for<n: int, l: loc> fn(&strg<l: i32[n]>) ensures l: i32[n + 1].

The name fhir is borrowed (pun intended) from rustc’s hir to refer to something a bit lower than the surface syntax.

Modules

lift

“Lift” HIR types into FHIR types.

visit

Structs

AliasReft

<qself as path>::name

AssocType

BaseTy

These are types of things that may be refined with indices or existentials

ConstArg

Crate

A map between rust definitions and flux annotations in their desugared fhir form.

EnumDef

Expr

FhirId

A unique identifier for a node in the AST. Like HirId it is composed of an owner and a local_id. We don’t generate ids for all nodes, but only for those we need to remember information elaborated during well-formedness checking to later be used during conversion into rty.

FieldDef

FnDecl

FnOutput

FnSig

FuncSort

GenericParam

GenericPredicates

Generics

Impl

ImplAssocReft

ImplItem

Item

ItemLocalId

An ItemLocalId uniquely identifies something within a given “item-like”.

MutTy

OpaqueTy

ParamId

PartialRes

See [rustc_hir::def::PartialRes]

Path

PathExpr

PathSegment

PolyFuncSort

PolyTraitRef

Qualifier

RefineArg

RefineParam

RefinedBy

Information about the refinement parameters associated with a type alias or an adt (struct/enum).

Requires

A predicate required to hold before calling a function.

SortDecl

SortPath

See flux_syntax::surface::SortPath

SpecFunc

StructDef

Trait

TraitAssocReft

TraitItem

Ty

TyAlias

TypeBinding

VariantDef

VariantRet

WhereBoundPredicate

VariantIdx

The source-order index of a variant in a type.

Enums

BaseTyKind

BinOp

ConstArgKind

Ensures

ExprKind

ExprRes

FluxItem

FluxLocalDefId

FluxOwnerId

Owner version of FluxLocalDefId

GenericArg

GenericBound

GenericParamKind

Ignored

A boolean used to mark whether a piece of code is ignored.

ImplItemKind

InferMode

Inference mode for a parameter.

ItemKind

Lifetime

Our surface syntax doesn’t have lifetimes. To deal with them we create a hole for every lifetime which we then resolve during annot_check when zipping against the lifted version.

Lit

Node

ParamKind

How the parameter was declared in the surface syntax. This is used to adjust how errors are reported and to control the inference mode.

PrimSort

QPath

RefineArgKind

Res

Sort

SortRes

SpecFuncKind

StructKind

TraitBoundModifier

TraitItemKind

Trusted

A boolean used to mark whether to mark wether code should be trusted.

TyKind

UnOp

Mutability

PrimTy

Not represented directly in the AST; referred to by name through a ty_path.

Type Aliases

Arena

GenericBounds

SortDecls