flux_middle::global_env

Struct GlobalEnv

pub struct GlobalEnv<'genv, 'tcx> {
    inner: &'genv GlobalEnvInner<'genv, 'tcx>,
}

Fields
inner: &'genv GlobalEnvInner<'genv, 'tcx>
Implementations

impl<'tcx> GlobalEnv<'_, 'tcx>

pub fn builtin_assoc_refts(self, def_id: DefId) -> Option<AssocRefinements>

pub fn builtin_assoc_reft_sort( self, assoc_id: FluxDefId, ) -> Option<EarlyBinder<FuncSort>>

pub fn builtin_assoc_reft_body( self, typing_env: TypingEnv<'tcx>, alias_reft: &AliasReft, ) -> Lambda

impl<'tcx> GlobalEnv<'_, 'tcx>

pub fn enter<'a, R>( tcx: TyCtxt<'tcx>, sess: &'a FluxSession, cstore: Box<CrateStoreDyn>, arena: &'a Arena, providers: Providers, f: impl for<'genv> FnOnce(GlobalEnv<'genv, 'tcx>) -> R, ) -> R

impl<'genv, 'tcx> GlobalEnv<'genv, 'tcx>

pub fn queried(self, def_id: DefId) -> bool

pub fn run_query_if_reached<K: DispatchKey, R>( self, key: K, query: impl FnOnce(Self, K) -> QueryResult<R>, ) -> QueryResult<R>

Runs a query only if the given key’s DefId was previously queried during checking.

During checking, we track all items transitively reached from explicitly included items. This method is used during metadata encoding to avoid triggering queries for items that were not reached. If the item was not previously queried, returns QueryErr::Ignored.

pub fn tcx(self) -> TyCtxt<'tcx>

pub fn sess(self) -> &'genv FluxSession

pub fn collect_specs(self) -> &'genv Specs

pub fn resolve_crate(self) -> &'genv ResolverOutput

pub fn lean_parent_dir(self) -> PathBuf

Parent directory of the Lean project.

pub fn temp_dir(self) -> &'genv TempDir

pub fn desugar(self, def_id: LocalDefId) -> QueryResult<Node<'genv>>

pub fn fhir_attr_map(self, def_id: LocalDefId) -> AttrMap<'genv>

pub fn fhir_crate(self) -> &'genv FluxItems<'genv>

pub fn alloc<T>(&self, val: T) -> &'genv T

pub fn alloc_slice<T: Copy>(self, slice: &[T]) -> &'genv [T]

pub fn alloc_slice_fill_iter<T, I>(self, it: I) -> &'genv [T]
where I: IntoIterator<Item = T>, I::IntoIter: ExactSizeIterator,

pub fn def_kind(&self, def_id: impl IntoQueryParam<DefId>) -> DefKind

pub fn alloc_slice_with_capacity<T, I>(self, cap: usize, it: I) -> &'genv [T]
where I: IntoIterator<Item = T>,

Allocates space to store cap elements of type T.

The elements are initialized using the supplied iterator. At most cap elements will be retrieved from the iterator. If the iterator yields fewer than cap elements, the returned slice will be of length less than the allocated capacity.

Panics

Panics if reserving space for the slice fails.

pub fn inlined_body(self, did: FluxDefId) -> Binder<Expr>

pub fn normalized_info(self, did: FluxDefId) -> FuncInfo

pub fn normalized_defns(self, krate: CrateNum) -> Rc<NormalizedDefns>

pub fn prim_rel_for(self, op: &BinOp) -> QueryResult<Option<&'genv PrimRel>>

pub fn qualifiers(self) -> QueryResult<&'genv [Qualifier]>

pub fn qualifiers_for( self, did: LocalDefId, ) -> QueryResult<impl Iterator<Item = &'genv Qualifier>>

Return all the qualifiers that apply to an item, including both global and local qualifiers.

pub fn reveals_for(self, did: LocalDefId) -> &'genv [FluxDefId]

Return the list of flux function definitions that should be revelaed for item

pub fn func_sort(self, def_id: impl IntoQueryParam<FluxDefId>) -> PolyFuncSort

pub fn func_span(self, def_id: impl IntoQueryParam<FluxDefId>) -> Span

pub fn should_inline_fun(self, def_id: FluxDefId) -> bool

pub fn variances_of(self, did: DefId) -> &'tcx [Variance]

pub fn mir(self, def_id: LocalDefId) -> QueryResult<Rc<BodyRoot<'tcx>>>

pub fn lower_generics_of( self, def_id: impl IntoQueryParam<DefId>, ) -> Generics<'tcx>

pub fn lower_predicates_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<GenericPredicates>

pub fn lower_type_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<EarlyBinder<Ty>>

pub fn lower_fn_sig( self, def_id: impl Into<DefId>, ) -> QueryResult<EarlyBinder<PolyFnSig>>

pub fn adt_def(self, def_id: impl IntoQueryParam<DefId>) -> QueryResult<AdtDef>

pub fn constant_info( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<ConstantInfo>

pub fn static_info( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<StaticInfo>

pub fn adt_sort_def_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<AdtSortDef>

pub fn sort_decl_param_count( self, def_id: impl IntoQueryParam<FluxDefId>, ) -> usize

pub fn check_wf(self, def_id: LocalDefId) -> QueryResult<Rc<WfckResults>>

pub fn impl_trait_ref( self, impl_id: DefId, ) -> QueryResult<EarlyBinder<TraitRef>>

pub fn generics_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<Generics>

pub fn refinement_generics_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<EarlyBinder<RefinementGenerics>>

pub fn predicates_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<EarlyBinder<GenericPredicates>>

pub fn assoc_refinements_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<AssocRefinements>

pub fn assoc_refinement(self, assoc_id: FluxDefId) -> QueryResult<AssocReft>

pub fn assoc_refinement_body_for_impl( self, trait_assoc_id: FluxDefId, impl_id: DefId, ) -> QueryResult<EarlyBinder<Lambda>>

Given the id of an associated refinement in a trait definition returns the body for the corresponding associated refinement in the implementation with id impl_id.

This function returns QueryErr::MissingAssocReft if the associated refinement is not found in the implementation and there’s no default body in the trait. This can happen if an extern spec adds an associated refinement without a default body because we are currently not checking compare_impl_item for those definitions.

pub fn default_assoc_refinement_body( self, trait_assoc_id: FluxDefId, ) -> QueryResult<Option<EarlyBinder<Lambda>>>

pub fn assoc_refinement_body( self, impl_assoc_id: FluxDefId, ) -> QueryResult<EarlyBinder<Lambda>>

pub fn sort_of_assoc_reft( self, assoc_id: FluxDefId, ) -> QueryResult<EarlyBinder<FuncSort>>

pub fn item_bounds( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<EarlyBinder<List<Clause>>>

pub fn type_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<EarlyBinder<TyOrCtor>>

pub fn fn_sig( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<EarlyBinder<PolyFnSig>>

pub fn variants_of( self, def_id: impl IntoQueryParam<DefId>, ) -> QueryResult<Opaqueness<EarlyBinder<PolyVariants>>>

pub fn variant_sig( self, def_id: DefId, variant_idx: VariantIdx, ) -> QueryResult<Opaqueness<EarlyBinder<PolyVariant>>>

pub fn lower_late_bound_vars( self, def_id: LocalDefId, ) -> QueryResult<List<BoundVariableKind>>

pub fn no_panic(self, def_id: impl IntoQueryParam<DefId>) -> bool

Whether the function is marked with #[flux::no_panic]

pub fn is_box(&self, res: Res) -> bool

pub fn def_id_to_param_index(&self, def_id: DefId) -> u32

pub(crate) fn cstore(self) -> &'genv CrateStoreDyn

pub fn has_trusted_impl(&self, def_id: DefId) -> bool

pub fn is_fn_output(&self, def_id: DefId) -> bool

The Output associated type is defined in FnOnce, and Fn/FnMut inherit it, so this should suffice to check if the def_id corresponds to LangItem::FnOnceOutput.

pub fn is_fn_call(&self, def_id: DefId) -> bool

Returns whether def_id is the call method in the Fn trait.

pub fn iter_local_def_id( self, ) -> impl Iterator<Item = LocalDefId> + use<'tcx, 'genv>

Iterator over all local def ids that are not a extern spec

pub fn iter_extern_def_id( self, ) -> impl Iterator<Item = DefId> + use<'tcx, 'genv>

pub fn maybe_extern_id(self, local_id: LocalDefId) -> MaybeExternId

pub fn resolve_id(self, def_id: DefId) -> ResolvedDefId

pub fn infer_opts(self, def_id: LocalDefId) -> InferOpts

pub fn trusted(self, def_id: LocalDefId) -> bool

Transitively follow the parent-chain of def_id to find the first containing item with an explicit #[flux::trusted(..)] annotation and return whether that item is trusted or not. If no explicit annotation is found, return false.

pub fn trusted_impl(self, def_id: LocalDefId) -> bool

pub fn is_dummy(self, def_id: LocalDefId) -> bool

Whether the item is a dummy item created by the extern spec macro.

See crate::Specs::dummy_extern

pub fn ignored(self, def_id: LocalDefId) -> bool

Transitively follow the parent-chain of def_id to find the first containing item with an explicit #[flux::ignore(..)] annotation and return whether that item is ignored or not. If no explicit annotation is found, return false.

pub fn should_fail(self, def_id: LocalDefId) -> bool

Whether the function is marked with #[flux::should_fail]

pub fn proven_externally(self, def_id: LocalDefId) -> Option<Span>

Whether the function is marked with #[proven_externally]

fn traverse_parents<T>( self, def_id: LocalDefId, f: impl FnMut(LocalDefId) -> Option<T>, ) -> Option<T>

Traverse the parent chain of def_id until the first node for which f returns Some.

impl<'genv, 'tcx> GlobalEnv<'genv, 'tcx>

pub fn fhir_iter_flux_items( self, ) -> impl Iterator<Item = (FluxLocalDefId, FluxItem<'genv>)>

pub fn fhir_sort_decl(&self, def_id: FluxLocalDefId) -> Option<&SortDecl>

pub fn fhir_spec_func_body( &self, def_id: FluxLocalDefId, ) -> Option<&'genv SpecFunc<'genv>>

pub fn fhir_qualifiers(self) -> impl Iterator<Item = &'genv Qualifier<'genv>>

pub fn fhir_primop_props(self) -> impl Iterator<Item = &'genv PrimOpProp<'genv>>

pub fn fhir_get_generics( self, def_id: LocalDefId, ) -> QueryResult<Option<&'genv Generics<'genv>>>

pub fn fhir_expect_refinement_kind( self, def_id: LocalDefId, ) -> QueryResult<&'genv RefinementKind<'genv>>

pub fn fhir_expect_item( self, def_id: LocalDefId, ) -> QueryResult<&'genv Item<'genv>>

pub fn fhir_expect_owner_node( self, def_id: LocalDefId, ) -> QueryResult<OwnerNode<'genv>>

pub fn fhir_node(self, def_id: LocalDefId) -> QueryResult<Node<'genv>>

impl GlobalEnv<'_, '_>

pub fn sort_of_self_ty_alias(self, alias_to: DefId) -> QueryResult<Option<Sort>>

pub fn sort_of_def_id(self, def_id: DefId) -> QueryResult<Option<Sort>>

fn sort_of_rust_ty(self, def_id: DefId, ty: Ty<'_>) -> QueryResult<Option<Sort>>

pub fn normalize_free_alias_sort(self, alias_ty: &AliasTy) -> QueryResult<Sort>