Struct flux_middle::global_env::GlobalEnv

pub struct GlobalEnv<'sess, 'tcx> {
    pub tcx: TyCtxt<'tcx>,
    pub sess: &'sess FluxSession,
    cstore: Box<CrateStoreDyn>,
    map: Map,
    queries: Queries<'tcx>,
}

Fields
tcx: TyCtxt<'tcx>sess: &'sess FluxSessioncstore: Box<CrateStoreDyn>map: Mapqueries: Queries<'tcx>
Implementations

impl<'sess, 'tcx> GlobalEnv<'sess, 'tcx>

pub fn new( tcx: TyCtxt<'tcx>, sess: &'sess FluxSession, cstore: Box<CrateStoreDyn> ) -> Self

pub fn providers(&mut self) -> &mut Providers

pub fn map(&self) -> &Map

pub fn map_mut(&mut self) -> &mut Map

pub fn defns(&self) -> QueryResult<&Defns>

pub fn qualifiers( &self, did: LocalDefId ) -> QueryResult<impl Iterator<Item = &Qualifier>>

pub fn func_decls(&self) -> impl Iterator<Item = &FuncDecl>

pub fn variances_of(&self, did: DefId) -> &[Variance]

pub fn mk_box(&self, ty: Ty, alloc: Ty) -> Ty

pub fn mir(&self, def_id: LocalDefId) -> QueryResult<Rc<Body<'tcx>>>

pub fn lower_type_of( &self, def_id: impl Into<DefId> ) -> QueryResult<EarlyBinder<Ty>>

pub fn lower_fn_sig(&self, def_id: DefId) -> QueryResult<EarlyBinder<PolyFnSig>>

pub fn adt_def(&self, def_id: impl Into<DefId>) -> QueryResult<AdtDef>

pub fn check_wf( &self, flux_id: impl Into<FluxLocalDefId> ) -> QueryResult<Rc<WfckResults>>

pub fn refparams_of( &self, def_id: impl Into<DefId> ) -> QueryResult<List<RefineParam>>

pub fn refparams_of_parent( &self, def_id: impl Into<DefId> ) -> QueryResult<List<RefineParam>>

pub fn generics_of(&self, def_id: impl Into<DefId>) -> QueryResult<Generics>

pub fn predicates_of( &self, def_id: DefId ) -> QueryResult<EarlyBinder<GenericPredicates>>

pub fn item_bounds( &self, def_id: DefId ) -> QueryResult<EarlyBinder<List<Clause>>>

pub fn type_of(&self, def_id: DefId) -> QueryResult<EarlyBinder<PolyTy>>

pub fn fn_sig( &self, def_id: impl Into<DefId> ) -> QueryResult<EarlyBinder<PolyFnSig>>

pub fn variants_of( &self, def_id: DefId ) -> QueryResult<Opaqueness<EarlyBinder<PolyVariants>>>

pub fn variant_sig( &self, def_id: DefId, variant_idx: VariantIdx ) -> QueryResult<Opaqueness<EarlyBinder<PolyVariant>>>

pub fn lower_late_bound_vars( &self, def_id: LocalDefId ) -> QueryResult<List<BoundVariableKind>>

pub fn get_generic_param(&self, def_id: LocalDefId) -> &GenericParam

pub fn is_box(&self, res: Res) -> bool

pub fn func_decl(&self, name: impl Borrow<Symbol>) -> Option<&FuncDecl>

pub fn const_by_name(&self, name: impl Borrow<Symbol>) -> Option<&ConstInfo>

pub fn sort_of_bty(&self, bty: &BaseTy) -> Option<Sort>

pub fn index_sorts_of(&self, def_id: impl Into<DefId>) -> &[Sort]

pub fn sort_of_res(&self, res: Res) -> Option<Sort>

pub fn early_bound_sorts_of(&self, def_id: DefId) -> &[Sort]

pub fn has_equality(&self, sort: &Sort) -> bool

Whether values of this sort can be compared for equality.

fn ctor_has_equality(&self, _: &SortCtor, args: &[Sort]) -> bool

For now all sort constructors have equality if all the generic arguments do. In the future we may have a more fine-grained notion of equality for sort constructors.

pub fn field_sort(&self, def_id: DefId, fld: Symbol) -> Option<&Sort>

pub fn field_index(&self, def_id: DefId, fld: Symbol) -> Option<usize>

pub fn refine_default_generic_args( &self, generics: &Generics, args: &GenericArgs ) -> QueryResult<GenericArgs>

pub fn refine_default( &self, generics: &Generics, rustc_ty: &Ty ) -> QueryResult<Ty>

pub fn refine_with_holes( &self, generics: &Generics, rustc_ty: &Ty ) -> QueryResult<Ty>

pub fn instantiate_arg_for_fun( &self, generics: &Generics, param: &GenericParamDef, arg: &GenericArg ) -> QueryResult<GenericArg>

pub fn instantiate_arg_for_constructor( &self, generics: &Generics, param: &GenericParamDef, arg: &GenericArg ) -> QueryResult<GenericArg>

pub(crate) fn cstore(&self) -> &CrateStoreDyn

pub fn hir(&self) -> Map<'tcx>

pub(crate) fn lookup_extern(&self, def_id: DefId) -> Option<DefId>

pub(crate) fn is_fn_once_output(&self, def_id: DefId) -> bool