Struct flux_middle::global_env::GlobalEnv

pub struct GlobalEnv<'genv, 'tcx> {
    inner: &'genv GlobalEnvInner<'genv, 'tcx>,
}

Fields
inner: &'genv GlobalEnvInner<'genv, 'tcx>
Implementations

impl<'tcx> GlobalEnv<'_, 'tcx>

pub fn enter<'a, R>( tcx: TyCtxt<'tcx>, sess: &'a FluxSession, cstore: Box<CrateStoreDyn>, arena: &'a Arena, providers: Providers, f: impl for<'genv> FnOnce(GlobalEnv<'genv, 'tcx>) -> R ) -> R

impl<'genv, 'tcx> GlobalEnv<'genv, 'tcx>

pub fn tcx(self) -> TyCtxt<'tcx>

pub fn hir(&self) -> Map<'tcx>

pub fn sess(self) -> &'genv FluxSession

pub fn collect_specs(self) -> &'genv Specs

pub fn resolve_crate(self) -> &'genv ResolverOutput

pub fn desugar(self, def_id: LocalDefId) -> QueryResult<Node<'genv>>

pub fn fhir_crate(self) -> &'genv Crate<'genv>

pub fn map(self) -> Map<'genv, 'tcx>

pub fn alloc<T>(&self, val: T) -> &'genv T

pub fn alloc_slice<T: Copy>(self, slice: &[T]) -> &'genv [T]

pub fn alloc_slice_fill_iter<T, I>(self, it: I) -> &'genv [T]
where I: IntoIterator<Item = T>, I::IntoIter: ExactSizeIterator,

pub fn def_kind(&self, def_id: impl Into<DefId>) -> DefKind

pub fn alloc_slice_with_capacity<T, I>(self, cap: usize, it: I) -> &'genv [T]
where I: IntoIterator<Item = T>,

Allocates space to store cap elements of type T.

The elements are initialized using the supplied iterator. At most cap elements will be retrived from the iterator. If the iterator yields fewer than cap elements, the returned slice will be of length less than the allocated capacity.

Panics

Panics if reserving space for the slice fails.

pub fn spec_func_defns(&self) -> QueryResult<&SpecFuncDefns>

pub fn qualifiers_for( self, did: LocalDefId ) -> QueryResult<impl Iterator<Item = &'genv Qualifier>>

Return all the qualifiers that apply to an item, including both global and local qualifiers.

pub fn func_decls( self ) -> QueryResult<impl Iterator<Item = &'genv SpecFuncDecl>>

pub fn func_decl(self, name: Symbol) -> QueryResult<SpecFuncDecl>

pub fn variances_of(self, did: DefId) -> &'tcx [Variance]

pub fn mk_box(&self, ty: Ty, alloc: Ty) -> Ty

pub fn mir(self, def_id: LocalDefId) -> QueryResult<Rc<Body<'tcx>>>

pub fn lower_generics_of( self, def_id: impl Into<DefId> ) -> QueryResult<Generics<'tcx>>

pub fn lower_predicates_of( self, def_id: impl Into<DefId> ) -> QueryResult<GenericPredicates>

pub fn lower_type_of( self, def_id: impl Into<DefId> ) -> QueryResult<EarlyBinder<Ty>>

pub fn lower_fn_sig(self, def_id: DefId) -> QueryResult<EarlyBinder<PolyFnSig>>

pub fn adt_def(self, def_id: impl Into<DefId>) -> QueryResult<AdtDef>

pub fn adt_sort_def_of( self, def_id: impl Into<DefId> ) -> QueryResult<AdtSortDef>

pub fn check_wf( self, flux_id: impl Into<FluxLocalDefId> ) -> QueryResult<Rc<WfckResults<'genv>>>

pub fn impl_trait_ref( self, impl_id: DefId ) -> QueryResult<Option<EarlyBinder<TraitRef>>>

pub fn generics_of(self, def_id: impl Into<DefId>) -> QueryResult<Generics>

pub fn refinement_generics_of( self, def_id: impl Into<DefId> ) -> QueryResult<RefinementGenerics>

pub fn predicates_of( self, def_id: impl Into<DefId> ) -> QueryResult<EarlyBinder<GenericPredicates>>

pub fn assoc_refinements_of( self, def_id: impl Into<DefId> ) -> QueryResult<AssocRefinements>

pub fn assoc_refinement_def( self, impl_id: DefId, name: Symbol ) -> QueryResult<EarlyBinder<Lambda>>

pub fn sort_of_assoc_reft( self, def_id: impl Into<DefId>, name: Symbol ) -> QueryResult<Option<EarlyBinder<FuncSort>>>

pub fn item_bounds( self, def_id: DefId ) -> QueryResult<EarlyBinder<List<Clause>>>

pub fn type_of(self, def_id: DefId) -> QueryResult<EarlyBinder<TyCtor>>

pub fn fn_sig( self, def_id: impl Into<DefId> ) -> QueryResult<EarlyBinder<PolyFnSig>>

pub fn variants_of( self, def_id: DefId ) -> QueryResult<Opaqueness<EarlyBinder<PolyVariants>>>

pub fn variant_sig( self, def_id: DefId, variant_idx: VariantIdx ) -> QueryResult<Opaqueness<EarlyBinder<PolyVariant>>>

pub fn lower_late_bound_vars( self, def_id: LocalDefId ) -> QueryResult<List<BoundVariableKind>>

pub fn get_generic_param( &self, def_id: LocalDefId ) -> QueryResult<&GenericParam<'_>>

pub fn is_box(&self, res: Res) -> bool

pub fn def_id_to_param_ty(&self, def_id: LocalDefId) -> ParamTy

pub fn def_id_to_param_index(&self, def_id: LocalDefId) -> u32

pub fn refine_default( self, generics: &Generics, rustc_ty: &Ty ) -> QueryResult<Ty>

pub fn refine_with_holes( self, generics: &Generics, rustc_ty: &Ty ) -> QueryResult<Ty>

pub(crate) fn cstore(self) -> &'genv CrateStoreDyn

pub fn is_fn_once_output(&self, def_id: DefId) -> bool

pub fn extern_id_of(self, def_id: DefId) -> QueryResult<Option<DefId>>

pub fn get_local_id_for_extern(self, extern_def_id: DefId) -> Option<LocalDefId>

pub fn ignores(self) -> &'genv UnordSet<IgnoreKey>

pub fn crate_config(self) -> Option<CrateConfig>

impl<'sess, 'tcx> GlobalEnv<'sess, 'tcx>

pub fn sort_of_alias_reft( self, alias: &AliasReft<'_> ) -> QueryResult<Option<FuncSort>>

pub fn sort_of_bty(self, bty: &BaseTy<'_>) -> QueryResult<Option<Sort>>

fn sort_of_path(self, path: &Path<'_>) -> QueryResult<Option<Sort>>

pub fn sort_of_self_ty_alias(self, alias_to: DefId) -> QueryResult<Option<Sort>>

fn sort_of_generic_param(self, def_id: LocalDefId) -> QueryResult<Option<Sort>>

fn sort_of_self_param(self, owner: DefId) -> QueryResult<Option<Sort>>

fn sort_of_generic_arg(self, arg: &GenericArg<'_>) -> QueryResult<Option<Sort>>

fn sort_of_ty(self, ty: &Ty<'_>) -> QueryResult<Option<Sort>>

fn sort_of_self_ty(self, def_id: DefId, ty: Ty<'_>) -> QueryResult<Option<Sort>>