Module flux_middle::rty::canonicalize

A canonical type is a type where all existentials and constraint predicates are hoisted to the top level. For example, the canonical version of (∃a. i32[a], ∃b. { i32[b] | b > 0}) is ∃a,b. { (i32[a], i32[b]) | b > 0}.

Canonicalization can be shallow or deep, by this we mean that some type constructors introduce new “scopes” that limit the hoisting. For instance, we are not allowed (in general) to hoist an existential type out of a generic argument, for example, in Vec<∃v. i32[v]> the existential inside the Vec cannot be hoisted out. However, the type inside the generic argument can be canonizalized locally inside the scope of the generic argument. Shallow canonicalization stops when finding type constructors. In contrast, deep canonicalization also canonizalizes inside type constructors. Note that some type constructors like shared references or boxes are transparent to hoisting and do not introduce a new scope.

It’s also important to note that canonizalization doesn’t imply any form of semantic equality and it is just a best effort to facilitate syntactic manipulation. For example, the types ∃a,b. (i32[a], i32[b]) and ∃a,b. (i32[b], i32[a]) are semantically equal but both are in canonical form (in the current implementation).

Modules

pretty 🔒

Structs

CanonicalConstrTy

ShallowHoister

Enums

CanonicalTy

A (shallowly) canonicalized type. This can be either of the form {T | p} or ∃v0,…,vn. {T | p}, where T doesnt have any (shallow) existential or constraint types.