flux_middle::rty::expr
Struct Expr

pub struct Expr {
    kind: Interned<ExprKind>,
    espan: Option<ESpan>,
}

Fields
kind: Interned<ExprKind>espan: Option<ESpan>
Implementations

impl Expr

pub fn at_opt(self, espan: Option<ESpan>) -> Expr

pub fn at(self, espan: ESpan) -> Expr

pub fn at_base(self, base: ESpan) -> Expr

pub fn span(&self) -> Option<ESpan>

pub fn tt() -> Expr

pub fn ff() -> Expr

pub fn and_from_iter(exprs: impl IntoIterator<Item = Expr>) -> Expr

pub fn or_from_iter(exprs: impl IntoIterator<Item = Expr>) -> Expr

pub fn and(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn or(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn zero() -> Expr

pub fn int_max(int_ty: IntTy) -> Expr

pub fn int_min(int_ty: IntTy) -> Expr

pub fn uint_max(uint_ty: UintTy) -> Expr

pub fn nu() -> Expr

pub fn is_nu(&self) -> bool

pub fn expect_adt(&self) -> (DefId, List<Expr>)

pub fn unit() -> Expr

pub fn var(var: Var) -> Expr

pub fn fvar(name: Name) -> Expr

pub fn evar(evar: EVar) -> Expr

pub fn bvar(debruijn: DebruijnIndex, var: BoundVar, kind: BoundReftKind) -> Expr

pub fn early_param(index: u32, name: Symbol) -> Expr

pub fn local(local: Local) -> Expr

pub fn constant(c: Constant) -> Expr

pub fn const_def_id(c: DefId) -> Expr

pub fn const_generic(param: ParamConst) -> Expr

pub fn aggregate(kind: AggregateKind, flds: List<Expr>) -> Expr

pub fn tuple(flds: List<Expr>) -> Expr

pub fn adt(def_id: DefId, flds: List<Expr>) -> Expr

pub fn from_bits(bty: &BaseTy, bits: u128) -> Expr

pub fn ite(p: impl Into<Expr>, e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn abs(lam: Lambda) -> Expr

pub fn hole(kind: HoleKind) -> Expr

pub fn kvar(kvar: KVar) -> Expr

pub fn alias(alias: AliasReft, args: List<Expr>) -> Expr

pub fn forall(expr: Binder<Expr>) -> Expr

pub fn binary_op(op: BinOp, e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn unit_adt(def_id: DefId) -> Expr

pub fn app(func: impl Into<Expr>, args: List<Expr>) -> Expr

pub fn global_func(func: Symbol, kind: SpecFuncKind) -> Expr

pub fn eq(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn unary_op(op: UnOp, e: impl Into<Expr>) -> Expr

pub fn ne(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn ge(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn gt(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn lt(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn le(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn implies(e1: impl Into<Expr>, e2: impl Into<Expr>) -> Expr

pub fn field_proj(e: impl Into<Expr>, proj: FieldProj) -> Expr

pub fn field_projs(e: impl Into<Expr>, projs: &[FieldProj]) -> Expr

pub fn path_proj(base: Expr, field: FieldIdx) -> Expr

pub fn not(&self) -> Expr

pub fn neg(&self) -> Expr

pub fn kind(&self) -> &ExprKind

pub fn is_atom(&self) -> bool

An expression is an atom if it is “self-delimiting”, i.e., it has a clear boundary when printed. This is used to avoid unnecessary parenthesis when pretty printing.

pub fn is_trivially_true(&self) -> bool

Simple syntactic check to see if the expression is a trivially true predicate. This is used mostly for filtering predicates when pretty printing but also to simplify types in general.

pub fn is_trivially_false(&self) -> bool

Simple syntactic check to see if the expression is a trivially false predicate.

fn is_true(&self) -> bool

Whether the expression is literally the constant true.

fn is_false(&self) -> bool

Whether the expression is literally the constant false.

pub fn from_const(tcx: TyCtxt<'_>, c: &Const) -> Expr

pub fn is_binary_op(&self) -> bool

fn const_op(op: &BinOp, c1: &Constant, c2: &Constant) -> Option<Constant>

pub fn simplify(&self) -> Expr

Simplify the expression by removing double negations, short-circuiting boolean connectives and doing constant folding. Note that we also have TypeFoldable::normalize which applies beta reductions for tuples and abstractions.

pub fn to_loc(&self) -> Option<Loc>

pub fn to_path(&self) -> Option<Path>

pub fn is_abs(&self) -> bool

pub fn is_unit(&self) -> bool

Whether this is an aggregate expression with no fields.

pub fn eta_expand_abs( &self, inputs: &BoundVariableKinds, output: Sort, ) -> Lambda

pub fn fold_sort(sort: &Sort, f: impl FnMut(&Sort) -> Expr) -> Expr

pub fn proj_and_reduce(&self, proj: FieldProj) -> Expr

Applies a field projection to an expression and optimistically try to beta reduce it

pub fn flatten_conjs(&self) -> Vec<&Expr>