Trait flux_middle::rty::fold::TypeFoldable

pub trait TypeFoldable: TypeVisitable {
    // Required method
    fn try_fold_with<F: FallibleTypeFolder>(
        &self,
        folder: &mut F
    ) -> Result<Self, F::Error>;

    // Provided methods
    fn fold_with<F: TypeFolder>(&self, folder: &mut F) -> Self { ... }
    fn normalize_projections<'tcx>(
        &self,
        genv: &GlobalEnv<'_, 'tcx>,
        infcx: &InferCtxt<'tcx>,
        callsite_def_id: DefId,
        refine_params: &[Expr]
    ) -> QueryResult<Self> { ... }
    fn normalize(&self, defns: &Defns) -> Self { ... }
    fn replace_holes(
        &self,
        f: impl FnMut(&[List<Sort>], HoleKind) -> Expr
    ) -> Self { ... }
    fn with_holes(&self) -> Self { ... }
    fn replace_evars(&self, evars: &EVarSol) -> Self { ... }
    fn shift_in_escaping(&self, amount: u32) -> Self { ... }
    fn shift_out_escaping(&self, amount: u32) -> Self { ... }
}

Required Methods

fn try_fold_with<F: FallibleTypeFolder>( &self, folder: &mut F ) -> Result<Self, F::Error>

Provided Methods

fn fold_with<F: TypeFolder>(&self, folder: &mut F) -> Self

fn normalize_projections<'tcx>( &self, genv: &GlobalEnv<'_, 'tcx>, infcx: &InferCtxt<'tcx>, callsite_def_id: DefId, refine_params: &[Expr] ) -> QueryResult<Self>

fn normalize(&self, defns: &Defns) -> Self

Normalize expressions by applying beta reductions for tuples and lambda abstractions.

fn replace_holes(&self, f: impl FnMut(&[List<Sort>], HoleKind) -> Expr) -> Self

Replaces all holes with the result of calling a closure. The closure takes a list with all the layers of bound variables at the point the hole was found. Each layer corresponds to the list of sorts bound at that level. The list is ordered from outermost to innermost binder, i.e., the last element is the binder closest to the hole.

fn with_holes(&self) -> Self

Turns each TyKind::Indexed into a TyKind::Exists with a TyKind::Constr and a hole. It also replaces all existing predicates with a hole. For example, Vec<{v. i32[v] | v > 0}>[n] becomes {n. Vec<{v. i32[v] | *}>[n] | *}.

fn replace_evars(&self, evars: &EVarSol) -> Self

fn shift_in_escaping(&self, amount: u32) -> Self

fn shift_out_escaping(&self, amount: u32) -> Self

Object Safety

This trait is not object safe.